chore(deps): bump project dependencies to latest versions
- [X] updated project dependencies, performed migrations;
- [X] verified project workability after
dependencies update;
- [X] updated README - added pipeline badges, and workspace tooling
references;
- [X] revised and updated package scripts;

BREAKING CHANGE: reinstall project dependencies, use `yarn install`
