+++ .eslintrc.js
@@ -12,7 +12,10 @@ module.exports = {
   parserOptions: {
     ecmaVersion: 2020,
     sourceType: 'module',
-    project: path.join(path.dirname(__filename), 'tsconfig.base.json'),
+    project: [
+      path.join(path.dirname(__filename), 'tsconfig.json'),
+      path.join(path.dirname(__filename), 'tsconfig.eslint-js.json'),
+    ],
   },
   extends: [
     'eslint:recommended',
@@ -401,6 +404,24 @@ module.exports = {
   },
 
   overrides: [
+    {
+      files: '**/*.js',
+      parser: 'esprima',
+      parserOptions: {
+        ecmaVersion: 2020,
+        sourceType: 'module',
+      },
+      extends: [
+        'eslint:recommended',
+        'plugin:prettier/recommended',
+        'plugin:@angular-eslint/recommended',
+      ],
+      plugins: [
+        'prettier',
+        '@angular-eslint', // https://github.com/angular-eslint/angular-eslint
+        'simple-import-sort', // https://github.com/lydell/eslint-plugin-simple-import-sort
+      ],
+    },
     {
       files: '**/main.ts',
       rules: {
