+++ apps/webaudio-spectrum-analyser-e2e/tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "types": ["cypress", "node"]
   },
-  "extends": "../../tsconfig.base.json",
+  "extends": "../../tsconfig.json",
   "files": [],
   "include": [],
   "references": [
