+++ apps/webaudio-spectrum-analyser/tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "types": ["node", "jest", "wasm", "util"]
   },
-  "extends": "../../tsconfig.base.json",
+  "extends": "../../tsconfig.json",
   "files": [],
   "include": [],
   "references": [
