+++ package.json
@@ -57,10 +57,11 @@
     "stylelint:libs:fix": "npx stylelint \"./libs/**/*.scss\" --fix",
     "test": "npx nx run-many --target=test --all --codeCoverage=false",
     "test:coverage": "npx nx run-many --target=test --all --codeCoverage=true",
-    "update": "ng update @nrwl/workspace",
-    "update:check": "ng update",
     "workspace:cleanup": "rm -rf ./documentation/* ./coverage/* ./dist/*",
-    "workspace:schematic": "nx workspace-schematic"
+    "workspace:schematic": "nx workspace-schematic",
+    "workspace:update:check": "ncu",
+    "workspace:update:cleanup": "rm -rf migrations.json",
+    "workspace:update:run-migrations": "nx migrate --run-migrations=migrations.json"
   },
   "husky": {
     "hooks": {
@@ -75,20 +76,20 @@
     "puppeteer_skip_chromium_download": true
   },
   "dependencies": {
-    "@angular/animations": "11.0.2",
-    "@angular/common": "11.0.2",
-    "@angular/compiler": "11.0.2",
-    "@angular/core": "11.0.2",
-    "@angular/forms": "11.0.2",
-    "@angular/platform-browser": "11.0.2",
-    "@angular/platform-browser-dynamic": "11.0.2",
-    "@angular/router": "11.0.2",
-    "@capacitor/android": "2.4.3",
+    "@angular/animations": "11.0.5",
+    "@angular/common": "11.0.5",
+    "@angular/compiler": "11.0.5",
+    "@angular/core": "11.0.5",
+    "@angular/forms": "11.0.5",
+    "@angular/platform-browser": "11.0.5",
+    "@angular/platform-browser-dynamic": "11.0.5",
+    "@angular/router": "11.0.5",
+    "@capacitor/android": "2.4.5",
     "@clr/angular": "3.1.5",
     "@clr/icons": "3.1.5",
     "@clr/ui": "3.1.5",
-    "@ngxs/store": "3.7.0",
-    "core-js": "3.7.0",
+    "@ngxs/store": "3.7.1",
+    "core-js": "3.8.1",
     "font-awesome": "4.7.0",
     "isomorphic-fetch": "3.0.0",
     "memo-decorator": "2.0.1",
@@ -100,61 +101,61 @@
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "10.0.1",
-    "@angular-devkit/build-angular": "0.1100.2",
-    "@angular-devkit/core": "11.0.2",
-    "@angular-devkit/schematics": "11.0.2",
-    "@angular-eslint/builder": "0.8.0-beta.1",
-    "@angular-eslint/eslint-plugin": "0.8.0-beta.1",
-    "@angular-eslint/eslint-plugin-template": "0.8.0-beta.1",
-    "@angular-eslint/template-parser": "0.8.0-beta.1",
-    "@angular/cli": "11.0.2",
-    "@angular/compiler-cli": "11.0.2",
-    "@angular/language-service": "11.0.2",
-    "@capacitor/cli": "2.4.3",
-    "@capacitor/core": "2.4.3",
+    "@angular-devkit/build-angular": "0.1100.5",
+    "@angular-devkit/core": "11.0.5",
+    "@angular-devkit/schematics": "11.0.5",
+    "@angular-eslint/builder": "0.8.0-beta.7",
+    "@angular-eslint/eslint-plugin": "0.8.0-beta.7",
+    "@angular-eslint/eslint-plugin-template": "0.8.0-beta.7",
+    "@angular-eslint/template-parser": "0.8.0-beta.7",
+    "@angular/cli": "11.0.5",
+    "@angular/compiler-cli": "11.0.5",
+    "@angular/language-service": "11.0.5",
+    "@capacitor/cli": "2.4.5",
+    "@capacitor/core": "2.4.5",
     "@commitlint/cli": "11.0.0",
     "@commitlint/config-conventional": "11.0.0",
-    "@ngxs/logger-plugin": "3.7.0",
-    "@ngxs/storage-plugin": "3.7.0",
-    "@nrwl/angular": "10.4.4",
-    "@nrwl/cli": "10.4.4",
-    "@nrwl/cypress": "10.4.4",
-    "@nrwl/eslint-plugin-nx": "10.4.4",
-    "@nrwl/jest": "10.4.4",
-    "@nrwl/node": "10.4.4",
-    "@nrwl/storybook": "10.4.4",
-    "@nrwl/tao": "10.4.4",
-    "@nrwl/workspace": "10.4.4",
-    "@storybook/addon-knobs": "6.1.2",
-    "@storybook/angular": "6.1.2",
-    "@types/jest": "26.0.15",
+    "@ngxs/logger-plugin": "3.7.1",
+    "@ngxs/storage-plugin": "3.7.1",
+    "@nrwl/angular": "11.0.18",
+    "@nrwl/cli": "11.0.18",
+    "@nrwl/cypress": "11.0.18",
+    "@nrwl/eslint-plugin-nx": "11.0.18",
+    "@nrwl/jest": "11.0.18",
+    "@nrwl/node": "11.0.18",
+    "@nrwl/storybook": "11.0.18",
+    "@nrwl/tao": "11.0.18",
+    "@nrwl/workspace": "11.0.18",
+    "@storybook/addon-knobs": "^6.1.11",
+    "@storybook/angular": "^6.1.11",
+    "@types/jest": "26.0.19",
     "@types/jsdom": "16.2.5",
-    "@types/node": "14.14.9",
-    "@typescript-eslint/eslint-plugin": "4.8.1",
-    "@typescript-eslint/parser": "4.8.1",
-    "babel-loader": "8.2.1",
-    "codelyzer": "6.0.0",
+    "@types/node": "14.14.16",
+    "@typescript-eslint/eslint-plugin": "4.11.0",
+    "@typescript-eslint/parser": "4.11.0",
+    "babel-loader": "8.2.2",
+    "codelyzer": "6.0.1",
     "commitizen": "4.2.2",
     "cpp-wasm-loader": "0.9.0",
-    "cypress": "5.6.0",
+    "cypress": "6.2.0",
     "cypress-multi-reporters": "1.4.0",
     "cz-conventional-changelog": "3.3.0",
-    "eslint": "7.14.0",
-    "eslint-config-prettier": "6.15.0",
-    "eslint-plugin-compat": "3.8.0",
+    "eslint": "7.16.0",
+    "eslint-config-prettier": "7.1.0",
+    "eslint-plugin-compat": "3.9.0",
     "eslint-plugin-cypress": "2.11.2",
-    "eslint-plugin-deprecation": "1.1.0",
-    "eslint-plugin-prettier": "3.1.4",
-    "eslint-plugin-rxjs": "2.1.4",
-    "eslint-plugin-simple-import-sort": "6.0.1",
+    "eslint-plugin-deprecation": "1.2.0",
+    "eslint-plugin-prettier": "3.3.0",
+    "eslint-plugin-rxjs": "2.1.5",
+    "eslint-plugin-simple-import-sort": "7.0.0",
     "eslint-watch": "7.0.0",
     "gh-pages": "3.1.0",
-    "husky": "4.3.0",
+    "husky": "4.3.6",
     "jest": "26.6.3",
     "jest-canvas-mock": "2.3.0",
     "jest-cpp-wasm": "0.1.0",
     "jest-preset-angular": "8.3.2",
-    "prettier": "2.2.0",
+    "prettier": "2.2.1",
     "prettier-stylelint": "0.4.2",
     "sort-json": "2.0.0",
     "sort-package-json": "1.48.0",
@@ -163,9 +164,9 @@
     "stylelint-config-rational-order": "0.1.2",
     "stylelint-prettier": "1.1.2",
     "ts-jest": "26.4.4",
-    "ts-node": "9.0.0",
+    "ts-node": "9.1.1",
     "typescript": "4.0.5",
-    "webpack-bundle-analyzer": "4.1.0"
+    "webpack-bundle-analyzer": "4.3.0"
   },
   "engines": {
     "node": ">=14.8.0",
