+++ tsconfig.eslint-js.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "allowJs": true
+  },
+  "include": ["./*.js", "./*.*.js", "./.*.js", "./.*.*.js"]
+}
